Fixed bug with positioning UniverseEntities with differing precision
before we weren't taking the different precision values of UniverseEntities into account when translating them.

+++ Assets/infinite-universe/CameraSpecificWorldScale.cs
@@ -33,7 +33,7 @@ public class CameraSpecificWorldScale : MonoBehaviour
         entity.transform.position = camera.transform.position;
         entity.transform.localScale = Vector3.zero;
       } else {
-        entity.transform.position = Vector3.Lerp(camera.transform.position, entity.transform.position*(float)entity.unitSize, (float)cameraScaling);
+        entity.transform.position = Vector3.Lerp(camera.transform.position, entity.transform.position*entity.unitSize, (float)cameraScaling);
         entity.transform.localScale *= (float)cameraUnitScaling;
       }
     }
